chore: refactor to remove some code duplication; create component container ABC
